1. Fixed the issue related to incorrect percentage during import for large amount of test cases 2. Updated the Jira field names from 'Data', 'Step', 'Result' to 'Test Data', 'Test Step', 'Expected Result' to have the same nomenclature as in Jira UI

+++ GITJIRAImporter/src/main/java/com/cigniti/util/Globalvars.java
@@ -41,7 +41,7 @@ public class Globalvars {
 	
 	public static String ExcelSheetPath = "";
 	public static String ExcelWorkSheetName = "";
-	public static int TotalTestCaseCount = 0;
+	public static double TotalTestCaseCount = 0;
 	public static int TotalTestCaseUploaded = 0;
 	
 	public static String strPreferencesPath = System.getenv("APPDATA") + "\\JiraImporter\\Preferences.config";
